recreate component as molecule
