+++ packages/components/molecules/f-restaurant-card/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 v0.1.0
 ------------------------------
-*September 30, 2021*
+*October 4, 2021*
 
 ### Added
-- `f-restaurant-card` component structure and config.
+- Add your change under the following headings: `Added`/`Changed`/`Deprecated`/`Removed`/`Fixed`/`Security`
