+++ packages/components/molecules/f-restaurant-card/package.json
@@ -1,13 +1,13 @@
 {
   "name": "@justeat/f-restaurant-card",
-  "description": "Fozzie Restaurant Card - Responsible for displaying restaurant data and acting as a link to a restaurant menu",
+  "description": "Fozzie Restaurant Card -  Responsible for displaying restaurant data and linking to a restaurant",
   "version": "0.1.0",
-  "main": "dist/f-restaurant-card.umd.min.js",
+  "main": "dist/f-restaurant-card.umd.min.js", 
   "files": [
     "dist",
-    "test-utils"
+    "test-utils" 
   ],
-  "homepage": "https://github.com/justeat/fozzie-components/tree/master/packages/components/organisms/f-restaurant-card",
+  "homepage": "https://github.com/justeat/fozzie-components/tree/master/packages/components/molecules/f-restaurant-card",
   "contributors": [
     "Github contributors <https://github.com/justeat/fozzie-components/graphs/contributors>"
   ],
@@ -31,10 +31,9 @@
     "lint": "vue-cli-service lint",
     "lint:fix": "yarn lint --fix",
     "lint:style": "vue-cli-service lint:style",
-    "report": "cd ../.. && yarn report",
     "test": "vue-cli-service test:unit",
-    "test-component:chrome": "wdio ../../../../wdio.conf.js --suite component",
-    "test-a11y:chrome": "wdio ../../../../wdio.conf.js --suite a11y"
+    "test-component:chrome": "cross-env-shell JE_ENV=local TEST_TYPE=component wdio ../../../../wdio-chrome.conf.js",
+    "test-a11y:chrome": "cross-env-shell JE_ENV=local TEST_TYPE=a11y wdio ../../../../wdio-chrome.conf.js --suite a11y" 
   },
   "browserslist": [
     "extends @justeat/browserslist-config-fozzie"
@@ -50,6 +49,7 @@
     "@vue/cli-plugin-babel": "4.4.6",
     "@vue/cli-plugin-eslint": "3.9.2",
     "@vue/cli-plugin-unit-jest": "4.4.6",
+    "@justeat/f-wdio-utils": "0.3.0",
     "@vue/test-utils": "1.0.3",
     "node-sass-magic-importer": "5.3.2"
   }
