+++ packages/components/molecules/f-restaurant-card/src/assets/scss/common.scss
