+++ packages/components/molecules/f-restaurant-card/test-utils/component-objects/f-restaurantCard-selectors.js
@@ -1,0 +1,1 @@
+export const COMPONENT = '[data-test-id="restaurantCard"]';
