+++ packages/components/molecules/f-restaurant-card/vue.config.js
