--- packages/components/organisms/f-restaurant-card/test-utils/component-objects/f-restaurantCard.component.js
@@ -1,5 +1,0 @@
-const restaurantCardComponent = () => $('[data-test-id="restaurantCard"]');
-
-exports.waitForRestaurantCardComponent = () => restaurantCardComponent().waitForExist();
-
-exports.isRestaurantCardComponentDisplayed = () => restaurantCardComponent().isDisplayed();
