--- packages/components/organisms/f-restaurant-card/test/specs/accessibility/axe-accessibility.spec.js
@@ -1,8 +1,0 @@
-import { getAccessibilityTestResults } from '../../../../../../../test/utils/axe-helper';
-
-describe('Accessibility tests', () => {
-    it('a11y - should test f-restaurantCard component WCAG compliance', () => {
-        // Act
-        const axeResults = getAccessibilityTestResults('f-restaurantCard');
-    });
-});
